pyproject.toml
[tool.poetry]
name = "llamaindex-starter-kit"
version = "0.1.0"
description = ""
authors = ["jalakoo <jalakoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
llama-index-readers-graphdb-cypher = "^0.1.3"
ipython = "^8.22.2"
llama-index-graph-stores-neo4j = "^0.1.3"
neo4j = "^5.18.0"
llama-index-llms-openai = "^0.1.12"
llama-index-embeddings-openai = "^0.1.6"
llama-index-tools-neo4j = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
